✨feat(navigation): add `fadeInOutPopTransitionSpec` and refactor package structure
- Added `fadeInOutPopTransitionSpec` with platform-specific implementations for Android and iOS.
- Updated `BaseNavDisplay` to use `fadeInOutPopTransitionSpec` as the default `predictivePopTransitionSpec`.
- Refactored package structure by moving files from `zone.ien.utils.navigation.data` and `zone.ien.utils.navigation.data.tracker` to `zone.ien.utils.navigation` and `zone.ien.utils.navigation.tracker`.
- ⬆️chore(lib): bumped version to 0.1.4-alpha06.
