+++ cmp-navigation/src/commonMain/kotlin/zone/ien/utils/navigation/ScreenInfo.kt
@@ -1,4 +1,4 @@
-package zone.ien.utils.navigation.data
+package zone.ien.utils.navigation
 
 import androidx.navigationevent.NavigationEventInfo
