+++ cmp-navigation/src/commonMain/kotlin/zone/ien/utils/navigation/Utils.kt
@@ -1,4 +1,4 @@
-package zone.ien.utils.navigation.data
+package zone.ien.utils.navigation
 
 import androidx.navigation3.runtime.NavBackStack
 import androidx.navigation3.runtime.NavKey
