+++ cmp-navigation/src/iosMain/kotlin/zone/ien/utils/navigation/transition/FadeInOutPopTransition.ios.kt
@@ -1,0 +1,9 @@
+package zone.ien.utils.navigation.transition
+
+import androidx.compose.animation.AnimatedContentTransitionScope
+import androidx.compose.animation.ContentTransform
+import androidx.navigation3.scene.Scene
+import androidx.navigation3.ui.defaultPredictivePopTransitionSpec
+import androidx.navigationevent.NavigationEvent
+
+actual fun <T : Any> fadeInOutPopTransitionSpec(): AnimatedContentTransitionScope<Scene<T>>.(@NavigationEvent.SwipeEdge Int) -> ContentTransform = defaultPredictivePopTransitionSpec()
