+++ gradle/libs.versions.toml
@@ -4,7 +4,7 @@ kotlin = "2.3.10"
 android-minSdk = "29"
 android-compileSdk = "36"
 android-targetSdk = "36"
-lib-version-name = "0.1.4-alpha05"
+lib-version-name = "0.1.4-alpha06"
 
 # pluginπ
 compose-plugin = "1.10.1"
